Combine WDIO tests
